% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vb.R
\name{e_log_pi}
\alias{e_log_pi}
\title{Compute Expectation of Log Initial Distribution}
\usage{
e_log_pi(alpha)
}
\arguments{
\item{alpha}{A torch tensor of shape (l,) containing the posterior
Dirichlet parameters for the initial distribution, where l = 2^k
is the number of latent classes}
}
\value{
A torch tensor of shape (l,) containing the expected values
of log(pi_l) for each latent class l
}
\description{
Computes the expectation of the logarithm of the initial distribution
parameters (pi) under a Dirichlet posterior distribution.
}
\details{
This function computes the expectation of log(pi_l) under a
Dirichlet posterior distribution. This expectation is used in
the variational lower bound computation.
}
\examples{
\dontrun{
alpha <- torch_tensor(c(2.5, 3.1, 4.2))
e_log_pi <- e_log_pi(alpha)
}

}
