% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%@\%}
\alias{\%@\%}
\title{Matrix Multiplication Operator for Torch Tensors}
\usage{
mat1 \%@\% mat2
}
\arguments{
\item{mat1}{A torch tensor (left operand)}

\item{mat2}{A torch tensor (right operand)}
}
\value{
A torch tensor resulting from the matrix multiplication of mat1 and
mat2
}
\description{
A custom infix operator that performs matrix multiplication between two torch
tensors. This is a convenience function that wraps the 'torch_matmul'
function.
}
\examples{
\dontrun{
a <- torch_randn(3, 4)
b <- torch_randn(4, 5)
result <- a \%@\% b
}

}
