% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{int_to_bin}
\alias{int_to_bin}
\title{Convert Integer to Binary Vector}
\usage{
int_to_bin(x, d)
}
\arguments{
\item{x}{An integer to convert (must be in range \link{0, 2^d - 1})}

\item{d}{The desired length of the binary vector}
}
\value{
A numeric vector of length d containing the binary representation
of x
}
\description{
Converts an integer to its binary representation as a vector of 0s and 1s
with a specified length.
}
\examples{
int_to_bin(5, 4)  # Returns c(1, 0, 1, 0)
int_to_bin(0, 3)  # Returns c(0, 0, 0)

}
