% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_alpha_trace}
\alias{plot_alpha_trace}
\title{Plot the trajectory of alpha parameters over iterations}
\usage{
plot_alpha_trace(
  alpha_trace,
  iter_trunc = 30,
  save = TRUE,
  path = "inst/figures/singlerun/",
  file_name = "alpha_trace.pdf"
)
}
\arguments{
\item{alpha_trace}{A matrix containing the alpha parameter traces}

\item{iter_trunc}{The number of iterations to truncate the plot (default: 30)}

\item{save}{Logical indicating whether to save the plot (default: TRUE)}

\item{path}{The directory path to save the plot (default: "inst/figures/singlerun/")}

\item{file_name}{The name of the file to save the plot (default: "alpha_trace.pdf")}
}
\value{
The ggplot object for the alpha trace plot
}
\description{
Plot the trajectory of alpha parameters over iterations
}
\examples{
alpha_trace <- matrix(rnorm(300), ncol = 3)
plot_alpha_trace(alpha_trace)

}
