% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_beta_recovery}
\alias{plot_beta_recovery}
\title{Plot the recovery plot of beta}
\usage{
plot_beta_recovery(
  beta_hat,
  beta_true,
  save = TRUE,
  path = "inst/figures/singlerun/",
  file_name = "beta_recovery.pdf"
)
}
\arguments{
\item{save}{Logical indicating whether to save the plot (default: TRUE)}

\item{path}{The directory path to save the plot (default: "inst/figures/singlerun/")}

\item{file_name}{The name of the file to save the plot (default: "pii_recovery.pdf")}

\item{pii_hat}{A vector containing the estimated pi values}

\item{pii_true}{A vector containing the true pi values}
}
\value{
The ggplot object for the pi recovery plot
}
\description{
Plot the recovery plot of beta
}
\examples{
pii_hat <- rnorm(100)
pii_true <- rnorm(100)
plot_pii_recovery(pii_hat, pii_true)

}
