% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_beta_trace}
\alias{plot_beta_trace}
\title{Plot the trajectory of beta parameters over iterations}
\usage{
plot_beta_trace(
  beta_trace,
  beta_true,
  iter_trunc = 30,
  save = TRUE,
  path = "inst/figures/singlerun/",
  file_name = "beta_trace.pdf"
)
}
\arguments{
\item{beta_trace}{A matrix containing the beta parameter traces}

\item{iter_trunc}{The number of iterations to truncate the plot (default: 30)}

\item{save}{Logical indicating whether to save the plot (default: TRUE)}

\item{path}{The directory path to save the plot (default: "inst/figures/singlerun/")}

\item{file_name}{The name of the file to save the plot (default: "beta_trace.pdf")}
}
\value{
The ggplot object for the beta trace plot
}
\description{
Plot the trajectory of beta parameters over iterations
}
\examples{
beta_trace <- matrix(rnorm(300), ncol = 3)
plot_beta_trace(beta_trace)

}
