% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_density}
\alias{plot_density}
\title{Plot the density of a vector}
\usage{
plot_density(
  vec,
  color = 2,
  xlab = "RMSE",
  save = TRUE,
  path = "",
  file_name = "rmse.pdf"
)
}
\arguments{
\item{vec}{A numeric vector for which to plot the density}

\item{color}{Color for the density plot (default: 2)}

\item{xlab}{Label for the x-axis (default: "RMSE")}

\item{save}{Logical indicating whether to save the plot (default: TRUE)}

\item{path}{The directory path to save the plot (default: "inst/figures/multiplerun/")}

\item{file_name}{The name of the file to save the plot (default: "rmse.pdf")}
}
\value{
A ggplot object for the density plot
}
\description{
Plot the density of a vector
}
\examples{
vec <- rnorm(100)
plot_density(vec)

}
