% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_elbo_trace}
\alias{plot_elbo_trace}
\title{Plot the trajectory of elbo over iterations}
\usage{
plot_elbo_trace(
  elbo_trace,
  iter_trunc = 30,
  save = TRUE,
  path = "inst/figures/singlerun/",
  file_name = "elbo_trace.pdf"
)
}
\arguments{
\item{elbo_trace}{A vector containing the elbo values}

\item{iter_trunc}{The number of iterations to truncate the plot (default: 30)}

\item{save}{Logical indicating whether to save the plot (default: TRUE)}

\item{path}{The directory path to save the plot (default: "inst/figures/singlerun/")}

\item{file_name}{The name of the file to save the plot (default: "elbo_trace.pdf")}
}
\value{
The ggplot object for the elbo trace plot
}
\description{
Plot the trajectory of elbo over iterations
}
\examples{
elbo_trace <- rnorm(100)
plot_elbo_trace(elbo_trace)

}
