% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_omega_trace}
\alias{plot_omega_trace}
\title{Plot the trajectory of omega parameters over iterations}
\usage{
plot_omega_trace(
  omega_trace,
  iter_trunc = 30,
  save = TRUE,
  path = "inst/figures/singlerun/",
  file_name = "omega_trace.pdf"
)
}
\arguments{
\item{omega_trace}{A matrix containing the omega parameter traces}

\item{iter_trunc}{The number of iterations to truncate the plot (default: 30)}

\item{save}{Logical indicating whether to save the plot (default: TRUE)}

\item{path}{The directory path to save the plot (default: "inst/figures/singlerun/")}

\item{file_name}{The name of the file to save the plot (default: "omega_trace.pdf")}
}
\value{
The ggplot object for the omega trace plot
}
\description{
Plot the trajectory of omega parameters over iterations
}
\examples{
omega_trace <- matrix(rnorm(300), ncol = 3)
plot_omega_trace(omega_trace)

}
