% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_tau_recovery}
\alias{plot_tau_recovery}
\title{Plot the recovery plot of tau}
\usage{
plot_tau_recovery(
  tau_hat,
  tau_true,
  save = TRUE,
  path = "inst/figures/singlerun/",
  file_name = "tau_recovery.pdf"
)
}
\arguments{
\item{tau_hat}{A vector containing the estimated tau values}

\item{tau_true}{A vector containing the true tau values}

\item{save}{Logical indicating whether to save the plot (default: TRUE)}

\item{path}{The directory path to save the plot (default: "inst/figures/singlerun/")}

\item{file_name}{The name of the file to save the plot (default: "tau_recovery.pdf")}
}
\value{
The ggplot object for the tau recovery plot
}
\description{
Plot the recovery plot of tau
}
\examples{
tau_hat <- rnorm(100)
tau_true <- rnorm(100)
plot_tau_recovery(tau_hat, tau_true)

}
