% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vb.R
\name{update_alpha}
\alias{update_alpha}
\title{Update Initial Distribution of Latent Attributes (Alpha)}
\usage{
update_alpha(alpha_prior, e_z)
}
\arguments{
\item{alpha_prior}{A torch tensor of shape (l,) containing the prior
parameters for the initial distribution, where l = 2^k is the number
of latent classes}

\item{e_z}{A torch tensor of shape (i, t, l) containing the expected
latent class memberships, where i is the number of individuals,
t is the number of time points, and l is the number of latent classes}
}
\value{
A torch tensor of shape (l,) containing the updated posterior
parameters for the initial distribution
}
\description{
Updates the variational posterior parameters for the initial distribution
(alpha) of latent attributes in the Hidden Markov Log-Linear Cognitive
Diagnostic Model using coordinate ascent variational inference.
}
\details{
This function implements the conjugate update for the Dirichlet
posterior of the initial latent class distribution by adding the expected
initial class counts (summed over individuals at the first time point)
to the prior parameters.
}
\examples{
\dontrun{
# Assuming appropriate torch tensors are available
alpha_updated <- update_alpha(alpha_prior, e_z)
}

}
