% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vb.R
\name{update_omega}
\alias{update_omega}
\title{Update Transition Probabilities (Omega)}
\usage{
update_omega(omega_prior, e_zz)
}
\arguments{
\item{omega_prior}{A torch tensor of shape (l, l) containing the prior
parameters for the transition probability matrix, where l = 2^k is the
number of latent classes}

\item{e_zz}{A torch tensor of shape (i, t-1, l, l) containing the expected
sufficient statistics for the transition counts, where i is the number
of individuals and t is the number of time points}
}
\value{
A torch tensor of shape (l, l) containing the updated posterior
parameters for the transition probability matrix
}
\description{
Updates the variational posterior parameters for the transition probabilities
(omega) in the Hidden Markov Log-Linear Cognitive Diagnostic Model using
coordinate ascent variational inference.
}
\details{
This function implements the conjugate update for the Dirichlet
posterior of the transition probabilities by adding the expected transition
counts (summed over individuals and time points) to the prior parameters.
}
\examples{
\dontrun{
# Assuming appropriate torch tensors are available
omega_updated <- update_omega(omega_prior, e_zz)
}

}
